Backend/requirements.txt
# Backend dependencies for Enterprise AI Assistant
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pypdf>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM: Groq (replaces Gemini)
groq>=0.4.0
